Extract test setup
