+++ test/planning-service.test.js
@@ -1,26 +1,6 @@
-const cds = require("@sap/cds/lib");
-const { GET, POST, PATCH, DELETE, axios, expect } = cds.test(__dirname + "/..");
-const EDIT = (url) => POST(`${url}/draftEdit`, {});
-const SAVE = (url) => POST(`${url}/draftActivate`);
-
-if (cds.User.default) {
-  cds.User.default = cds.User.Privileged;
-} else {
-  cds.User = cds.User.Privileged;
-}
+const { GET, POST, PATCH, DELETE, EDIT, SAVE, expect } = require("./utils");
 
 describe("Planning service", () => {
-  let validateStatus;
-
-  beforeAll(() => {
-    validateStatus = axios.defaults.validateStatus;
-    axios.defaults.validateStatus = () => true;
-  });
-
-  afterAll(() => {
-    axios.defaults.validateStatus = validateStatus;
-  });
-
   it("returns customers with name", async () => {
     const { status, data } =
       await GET`/planning/Customers?$select=formattedName`;
